backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.7
spacy==3.7.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
numpy==1.24.4
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5